[feature]: support singing voice conversion (#48)
