+++ .gitignore
@@ -3,9 +3,10 @@
 # Edit at https://www.gitignore.io/?templates=linux,python,windows,pycharm+all,visualstudiocode
 
 ### Backend ###
-# japanese
+# large files
 *.pth
 *.pkl
+*.pt
 mecab-0.996.tar.gz
 
 ### Frontend ###
