+++ app/backend/requirements.txt
@@ -9,3 +9,6 @@ scipy>=1.11.2
 torch>=2.0.1
 tqdm>=4.66.1
 Unidecode>=1.3.6
+fairseq
+torchaudio
+SoundFile
